Add files via upload
Concepts tested by the program:

1. Work with random numbers and random number generation

2. Work with arrays

3. Use file operations

4. To learn and use Bubble sort to sort an array

5. To learn and use binary search to a sorted array

6. To learn and use sequential search to an array

7. Implement functions besides function main()

Program: Write, compile, and run a C++ program that generates 200 random

integers in between 1 to 200. Some of the randomly generated integers may be

repeated as well. Next sort the randomly generated 200 integers using the Bubble

sort algorithm. Store the sorted numbers in an output data file. Then prompt the

user to enter an integer in between 1 to 200 (inclusive) out of the keyboard as a

key element. Next find out the location of the key element in the sorted array using

both the binary search and the sequential search algorithms. The result returned by

binary search and the sequential search algorithm may vary for